src/RAG/requirements.txt
fastapi
uvicorn
huggingface_hub
psycopg2-binary
sentence-transformers
faiss-cpu
pydantic
python-dotenv